Renamed main directory to app
