+++ app/biomedqa/extractors/loader.py
@@ -2,13 +2,13 @@
     Script to load data into Elasticsearch
 """
 import json
-from elasticsearchclient import ElasticsearchClient
+from app.biomedqa.retrievers.elasticsearchclient import ElasticsearchClient
 
-connection_config = json.load(open("retriever/config.json", encoding="utf-8"))
+connection_config = json.load(open("config.json", encoding="utf-8"))
 
 conn = ElasticsearchClient(connection_config)
 
 # Create Elasticsearch Index named pubmed
-conn.create("pubmed2", "retriever/schema.json")
+conn.create("pubmed2", "schema.json")
 
 conn.load("retriever/data.json")
