+++ app/templates/include/searchbar.html
